resolves #64 add an API to manipulate the CLI
